+++ src/main/java/by/intereson/ebooksservice/controllers/book/ReadBookStartController.java
@@ -1,4 +1,4 @@
-package by.intereson.ebooksservice.controllers;
+package by.intereson.ebooksservice.controllers.book;
 
 import by.intereson.ebooksservice.entities.Book;
 import by.intereson.ebooksservice.services.BookService;
@@ -9,21 +9,26 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.util.List;
 
 import static by.intereson.ebooksservice.utils.Constants.*;
 
-@WebServlet(urlPatterns = EXIT_URL)
-public class ExitController extends HttpServlet {
+@WebServlet(urlPatterns = MAIN_PAGE_URL)
+public class ReadBookStartController extends HttpServlet {
+
+
     private final BookService bookService = BookServiceImpl.getInstance();
+
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        HttpSession session = req.getSession();
-        session.invalidate();
         List<Book> books = bookService.readBooks();
         req.setAttribute(BOOKS, books);
-        req.getRequestDispatcher(MAIN_PAGE_URL).forward(req,resp);
+        req.getRequestDispatcher(BOOKS_FOR_USER_PAGE).forward(req, resp);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        doGet(req,resp);
     }
 }
