+++ src/main/java/by/intereson/ebooksservice/controllers/user/CreateUserController.java
@@ -1,4 +1,4 @@
-package by.intereson.ebooksservice.controllers;
+package by.intereson.ebooksservice.controllers.user;
 
 import by.intereson.ebooksservice.entities.Book;
 import by.intereson.ebooksservice.entities.User;
