+++ src/main/java/by/intereson/ebooksservice/entities/User.java
@@ -1,5 +1,6 @@
 package by.intereson.ebooksservice.entities;
 
+import by.intereson.ebooksservice.enums.UserType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
