+++ src/main/webapp/pages/errors/ErrorLogin.jsp
@@ -1,4 +1,5 @@
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
+<%@ page contentType="text/html;charset=UTF-8"  %>
 <html>
 <head>
     <title>Error</title>
@@ -7,10 +8,10 @@
 <p>Error Login</p>
 <p>The entered login or email already exists</p>
 <p>Try again</p>
-<a href="/users/authentication">
+<a href="<c:url value="/users/authentication"/>">
   <p>Enter</p>
 </a>
-<a href="/users/registration">
+<a href="<c:url value="/users/registration"/>">
     <p>Registration</p>
 </a>
 </body>
