+++ src/main/webapp/pages/errors/errorAccess.jsp
@@ -1,4 +1,5 @@
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
+<%@ page contentType="text/html;charset=UTF-8"  %>
 <html>
 <head>
     <title>Error</title>
@@ -6,10 +7,10 @@
 <body>
 <p>Error Access</p>
 
-<a href="/users/authentication">
+<a href="<c:url value="/users/authentication"/>">
     <p>Enter</p>
 </a>
-<a href="/users/registration">
+<a href="<c:url value="/users/registration"/>">
     <p>Registration</p>
 </a>
 </body>
