+++ src/main/webapp/pages/errors/errorDataBookPage.jsp
@@ -1,11 +1,12 @@
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
+<%@ page contentType="text/html;charset=UTF-8"  %>
 <html>
 <head>
     <title>Error</title>
 </head>
 <body>
 <p>Error Data</p>
-<a href="/books/create">
+<a href="<c:url value="/books/create"/>">
     <p>Try again</p>
 </a>
 </body>
