+++ src/main/webapp/pages/users/deleteUser.jsp
@@ -1,4 +1,4 @@
-<%@ page contentType="text/html;charset=UTF-8" language="java" isELIgnored="false" %>
+<%@ page contentType="text/html;charset=UTF-8"  isELIgnored="false" %>
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <html>
 <head>
@@ -20,8 +20,9 @@
     </tr>
     </thead>
     <tbody>
+    <jsp:useBean id="users" scope="request" type="java.util.List"/>
     <c:forEach var="user" items="${users}">
-        <form action="/users/delete" method="post">
+        <form action="<c:url value="/users/delete"/>" method="post">
             <tr>
                 <td> ${user.id} </td>
                 <td>${user.name}</td>
@@ -39,7 +40,7 @@
                     <input type="submit" value="DELETE">
                     <input type="hidden" name="id" value="${user.id}">
                 </td>
-                <a href="/users/read">  <p>CANCEL</p> </a>
+                <a href="<c:url value="/users/read"/>">  <p>CANCEL</p> </a>
             </tr>
         </form>
     </c:forEach>
